#  Convert label values to strings in base Metric class