Currently, you get an inconsistent experience when using different store
implementations (e.g. `SingleThreaded` vs `DirectFileStore`).

Because `DirectFileStore` has to write your metrics into a file, it
has to convert label values into strings. That means that when you
access them they're strings. Other stores don't do this, so you can
access them as symbols if you provided them as symbols.

To give users a consistent experience when swapping between stores, this
commit changes the Metric base class to convert the label values to
strings before they get anywhere near the underlying store.

current master:
```
            0 labels      3.732M (± 1.9%) i/s -     18.662M in   5.003057s
            2 labels      1.008M (± 2.4%) i/s -      5.039M in   5.000649s
          100 labels     31.093k (± 2.2%) i/s -    156.208k in   5.026452s
  2 lab, half cached      1.017M (± 2.3%) i/s -      5.097M in   5.013142s
100 lab, half cached     32.958k (± 2.6%) i/s -    167.076k in   5.072940s
   2 lab, all cached      3.788M (± 1.5%) i/s -     18.959M in   5.006396s
 100 lab, all cached      3.772M (± 1.9%) i/s -     18.873M in   5.005481s
```

this branch:
```
            0 labels      3.689M (± 1.3%) i/s -     18.450M in   5.002189s
            2 labels    713.961k (± 2.7%) i/s -      3.596M in   5.039963s
          100 labels     21.414k (± 1.9%) i/s -    107.378k in   5.016144s
  2 lab, half cached    791.084k (± 2.3%) i/s -      3.968M in   5.017964s
100 lab, half cached     26.202k (± 3.2%) i/s -    132.243k in   5.052325s
   2 lab, all cached      3.724M (± 1.9%) i/s -     18.718M in   5.028441s
 100 lab, all cached      3.712M (± 1.1%) i/s -     18.723M in   5.043903s
```

master before today's performance improvements:
```
            0 labels      3.712M (± 1.8%) i/s -     18.726M in   5.046030s
            2 labels    645.876k (± 4.0%) i/s -      3.251M in   5.043007s
          100 labels     18.990k (± 3.9%) i/s -     95.676k in   5.046566s
  2 lab, half cached    668.194k (± 3.3%) i/s -      3.361M in   5.036575s
100 lab, half cached     20.290k (± 2.3%) i/s -    101.796k in   5.020072s
   2 lab, all cached      3.731M (± 2.0%) i/s -     18.831M in   5.049451s
 100 lab, all cached      3.747M (± 3.4%) i/s -     18.718M in   5.003039s
```

Fixes #115